4.25 - 4.19

+++ part4/controllers/blogs.js
@@ -1,10 +1,23 @@
 const express = require("express");
 const blogRouter = express.Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+
+const getTokenFrom = (request) => {
+  const authorization = request.get("authorization");
+  if (authorization && authorization.startsWith("Bearer ")) {
+    return authorization.replace("Bearer ", "");
+  }
+  return null;
+};
 
 blogRouter.get("/", async (request, response, next) => {
   try {
-    const blogs = await Blog.find({});
+    const blogs = await Blog.find({}).populate("user", {
+      username: 1,
+      name: 1,
+    });
     response.json(blogs);
   } catch (error) {
     next(error);
@@ -13,7 +26,17 @@ blogRouter.get("/", async (request, response, next) => {
 
 blogRouter.post("/", async (request, response, next) => {
   try {
-    const blog = await new Blog(request.body).save();
+    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+    const user = await User.findById(decodedToken.id);
+    const payload = { ...request.body, user: user.id };
+    const blog = await new Blog(payload).save();
+
+    user.blogs = user.blogs.concat(blog._id);
+    await user.save();
+
     response.status(201).json(blog);
   } catch (error) {
     next(error);
@@ -22,7 +45,16 @@ blogRouter.post("/", async (request, response, next) => {
 
 blogRouter.delete("/:id", async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+    const user = await User.findById(decodedToken.id);
     await Blog.findByIdAndDelete(request.params.id);
+
+    user.blogs = user.blogs.filter((blog) => blog._id !== request.params.id);
+    await user.save();
+
     response.status(204).send();
   } catch (error) {
     next(error);
@@ -31,10 +63,28 @@ blogRouter.delete("/:id", async (request, response, next) => {
 
 blogRouter.patch("/:id", async (request, response, next) => {
   try {
-    const blog = await Blog.findByIdAndUpdate(request.params.id, request.body, {
-      new: true,
+    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+    const user = await User.findById(decodedToken.id);
+
+    const updatedBlog = await Blog.findByIdAndUpdate(
+      request.params.id,
+      request.body,
+      {
+        new: true,
+      }
+    );
+    user.blogs = user.blogs.map((blog) => {
+      if (blog._id === request.params.id) {
+        return updatedBlog;
+      } else {
+        return blog;
+      }
     });
-    response.status(200).json(blog);
+    await user.save();
+    response.status(200).json(updatedBlog);
   } catch (error) {
     next(error);
   }
